+++ src/main/java/com/nofirst/zhihu/service/AnswerService.java
@@ -13,7 +13,7 @@ import com.nofirst.zhihu.security.AccountUser;
 public interface AnswerService {
     Answer show(Long id);
 
-    void store(Long questionId, AnswerDto answerDto, AccountUser accountUser);
+    void store(Integer questionId, AnswerDto answerDto, AccountUser accountUser);
 
     void markAsBest(Long answerId, AccountUser accountUser);
