+++ src/test/java/com/nofirst/zhihu/controller/PublishQuestionTest.java
@@ -78,7 +78,7 @@ class PublishQuestionTest {
         QuestionExample example = new QuestionExample();
         QuestionExample.Criteria criteria = example.createCriteria();
         criteria.andPublishedAtIsNotNull();
-        int beforeCount = questionMapper.countByExample(example);
+        long beforeCount = questionMapper.countByExample(example);
         // when
         this.mockMvc.perform(post("/questions/{questionId}/published-questions", question.getId())
                         .contentType(MediaType.APPLICATION_JSON)
@@ -88,7 +88,7 @@ class PublishQuestionTest {
                 .andExpect(jsonPath("$.code").value(ResultCode.SUCCESS.getCode()));
 
         // then
-        int afterCount = questionMapper.countByExample(example);
+        long afterCount = questionMapper.countByExample(example);
         // 调用之后 question 增加了 1 条
         assertThat(afterCount - beforeCount).isEqualTo(1);
     }
@@ -97,7 +97,7 @@ class PublishQuestionTest {
     void guests_may_not_publish_questions() throws Exception {
         // given
         Question question = QuestionFactory.createUnpublishedQuestion();
-        question.setId(1L);
+        question.setId(1);
         // when
         this.mockMvc.perform(post("/questions/{questionId}/published-questions", question.getId())
                         .contentType(MediaType.APPLICATION_JSON)
